+++ models/fan-controller.go
@@ -17,7 +17,7 @@ type FanController struct {
 	gorm.Model
 	 Name string `json:"name" form:"name"`
 	 Price float64 `json:"price" form:"price"`
-	 Channels int `json:"channels" form:"channels"`
+	 Channels float64 `json:"channels" form:"channels"`
 	 ChannelWattage string `json:"channel_wattage" form:"channel_wattage"`
 	 Pwm bool `json:"pwm" form:"pwm"`
 	 FormFactor string `json:"form_factor" form:"form_factor"`
@@ -48,7 +48,7 @@ func (FanController)ImportData(db *gorm.DB) error {
 		m := FanController{
 			Name: records[i][0],
 			Price: utils.ToFloat64(records[i][1]),
-			Channels: utils.ToInt(records[i][2]),
+			Channels: utils.ToFloat64(records[i][2]),
 			ChannelWattage: records[i][3],
 			Pwm: utils.ToBool(records[i][4]),
 			FormFactor: records[i][5],
