+++ models/optical-drive.go
@@ -17,9 +17,9 @@ type OpticalDrive struct {
 	gorm.Model
 	 Name string `json:"name" form:"name"`
 	 Price float64 `json:"price" form:"price"`
-	 Bd int `json:"bd" form:"bd"`
-	 Dvd int `json:"dvd" form:"dvd"`
-	 Cd int `json:"cd" form:"cd"`
+	 Bd float64 `json:"bd" form:"bd"`
+	 Dvd float64 `json:"dvd" form:"dvd"`
+	 Cd float64 `json:"cd" form:"cd"`
 	 BdWrite string `json:"bd_write" form:"bd_write"`
 	 DvdWrite string `json:"dvd_write" form:"dvd_write"`
 	 CdWrite string `json:"cd_write" form:"cd_write"`
@@ -49,9 +49,9 @@ func (OpticalDrive)ImportData(db *gorm.DB) error {
 		m := OpticalDrive{
 			Name: records[i][0],
 			Price: utils.ToFloat64(records[i][1]),
-			Bd: utils.ToInt(records[i][2]),
-			Dvd: utils.ToInt(records[i][3]),
-			Cd: utils.ToInt(records[i][4]),
+			Bd: utils.ToFloat64(records[i][2]),
+			Dvd: utils.ToFloat64(records[i][3]),
+			Cd: utils.ToFloat64(records[i][4]),
 			BdWrite: records[i][5],
 			DvdWrite: records[i][6],
 			CdWrite: records[i][7],
